requirements.txt
# Core web app dependencies
Flask==3.0.0
groq==0.29.0
python-dotenv==1.1.1
requests==2.32.4
beautifulsoup4==4.13.4
PyPDF2==3.0.1
pdfplumber==0.7.6

# Required by dependencies
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
regex==2024.11.6
sniffio==1.3.1
soupsieve==2.7
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.7.0
itsdangerous==2.1.2